fixing settings load bugs, added tracing
### XmlSettingManager
-  **fixed always loading setting default value:**
manager was giving the property node instead of the value node to converters, resulting in them not being able to parse the data and returning null.
- **added clause to not try to convert `Type` objects (since they do need to be converted to be assigned)**
- **fixed `Reset`:**
document was getting cleared but `XmlDocument` needs to have a root node to be a valid XML file. Now after document is getting cleared, a new "Settings" root node is added
- **added enum support (without converters)**
- **added tracing information**
  - if the node for this class in `null`
  - if the manager fails to load a property value from the XML file

### Converters
#### SizeSettingConverter
- using `nameof` to create attributes

#### From changes in `XmlSettingManager` (both classes)
- checking node name in `Restore` for conversion
- kept null check in restore as before